Fix editUrl in docusaurus config (#1)
After having a quick look at the new site, I saw that the links to edit
both the docs and blog section were taking me over to a 404 in the
GitHub repository. There was an extra "docs" and "blog" in the URLs that
were being used.

I actually have no idea whether this change will work, but based on what
I am seeing being generated, it would appear to make sense.